Partial integration of some OpenProject models

+++ src/main/java/com/github/jpmand/openproject/client/api/models/OPRoleModel.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 public class OPRoleModel extends OPBaseResource {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("name")
     private String name;
@@ -19,16 +19,16 @@ public class OPRoleModel extends OPBaseResource {
     public OPRoleModel() {
     }
 
-    public OPRoleModel(Integer id, String name) {
+    public OPRoleModel(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
